HUGE refactoring
- rinominato tutti i file .c in .cpp (cosi' compila nativamente in c++)

- rinominato tutti i file .h in .hpp (idem come sopra)

- riformattato tutto il codice con astyle (non avete idea di quanto
semplifica i diff avere il codice sempre formtattato nello stesso modo)
messa la configurazione di astyle e astyle nel repo, e aggiunta la
riformattazione allo step di build

- configurato cmake per il build (avevo provato autotools ma i sono
confermato nella mia idea: fa recere). Ad ogni compilazione viene
ricreato il file release.hpp in modo da avere la data di compilazione
sempre aggiornata

- aggiunto uno script (build.sh) che genera i file locali (config.hpp e
release.hpp) e compila il codice "a pulito". NON toccatel senza
avvisarmi perché è lo stesso che lanciano i deploy sul server, quindi è
bene che faccia sempre le stesse cose

- ho pulito anche la macchina vagrant e aggiunto il lancio di build.sh
alla fine del provisioning. Vi consiglio un bel vagrant destroy &&
vagrant up o in alternativa vagrant provisio per riconfigurarla

- per finire in bellezza ho impostato lo standard del compilatore a c11

+++ build.sh
@@ -1,12 +1,26 @@
 #!/bin/bash
 # Utilizzato dal sistema di deploy.
-# In ambiente di sviluppo va lancianto solo se si aggiunge un nuovo file
+# In ambiente di sviluppo va lanciato solo se si aggiunge un nuovo file
 cd $(dirname "$0")
-rm -f mudroot/myst
-rm -f  CMakeCache.txt
-cmake .
-make
-if [ -x src/myst ] ; then
-	cp src/myst mudroot/myst
+environment=${1:-vagrant}
+conf="$HOME/Confs/$environment.conf"
+if [ -f $conf ] ; then
+	echo '#ifndef CONFIG_HPP_' >src/config.hpp.in
+	echo '#define CONFIG_HPP_' >>src/config.hpp.in
+	cat  $HOME/Confs/$environment.conf | sed -e 's/#/\/\//' | sed -e 's/=/ /' | sed -e 's/\(.*\)/#define \1/' >>src/config.hpp.in
+	echo '#endif /*CONFIG_HPP_*/' >>src/config.hpp.in
+	cat src/config.hpp.in
+	rm -f src/release.h
+	rm -f mudroot/myst
+	rm -f  CMakeCache.txt
+	cmake .
+	make clean && make
+	if [ -x src/myst ] ; then
+		cp src/myst mudroot/myst
+		exit 0
+	fi
+	exit 1
+else
+	exit 255
 fi
